requirements.txt
torch==2.2.1
torchvision==0.17.1
mlflow==2.11.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
boto3==1.34.58
